Move index based methods to separate directory
