+++ services/blockchain-indexer/src/blocks.js
@@ -14,14 +14,14 @@
  *
  */
 const BluebirdPromise = require('bluebird');
-const { getIndexedAccountInfo } = require('../utils/accountUtils');
-const { getTxnMinFee } = require('../utils/transactionsUtils');
-const { parseToJSONCompatObj } = require('../utils/jsonTools');
+const { getIndexedAccountInfo } = require('./utils/accountUtils');
+const { getTxnMinFee } = require('./utils/transactionsUtils');
+const { parseToJSONCompatObj } = require('./utils/jsonTools');
 
 const { getCurrentHeight } = require('./constants');
 
-const blockchainStore = require('./blockchainStore');
-const { requestRpc } = require('../utils/appContext');
+const blockchainStore = require('./indexer/blockchainStore');
+const { requestRpc } = require('./utils/appContext');
 
 // Genesis height can be greater that 0
 // Blockchain starts form a non-zero block height
